requirements.txt
pbr>=1.6
pyOpenSSL>=0.15
cffi
keystone>=9.0.0,<10.0.0
oslo.config>=2.3.0
oslo.log>=1.8.0
oslo.serialization>=1.4.0
